refactor sync
